Informe de avance – Ajuste multiempresa, multisede y multipdv en Ventas Manuales y POS
El día de hoy se realizaron ajustes enfocados en que los módulos de Ventas Manuales y POS trabajen con el contexto real de acceso del usuario, evitando que el sistema tome por defecto la primera empresa o el primer PDV encontrado.

1. Validación del acceso por PDV seleccionado
Se confirmó y reforzó la lógica de login para que el usuario ingrese con un PDV específico, validando que:

el PDV exista,
la sede exista,
la empresa del PDV coincida con la empresa del usuario,
y que el usuario tenga autorización para ingresar a ese PDV.

2. Claims de sesión para contexto operativo
Se dejó definido el contexto de sesión con los claims:

EmpresaId
PdvId
SedeId
además de nombre de empresa, nombre de sede y nombre del PDV

Con esto, el sistema ya tiene identificado desde login en qué empresa, sede y punto de venta quedó trabajando el usuario.

3. Contexto de acceso en Ventas Manuales
En PedidoController se organizó el uso de un ContextoAccesoDto para leer directamente desde los claims la:

cédula,
empresa actual,
sede actual,
PDV actual,
rol del usuario.

Esto permite que las consultas y procesos de ventas manuales ya no dependan de búsquedas ambiguas del usuario, sino del contexto real con el que inició sesión.

4. Filtrado de ventas y facturas por contexto actual
Se ajustó la lógica para que los listados de ventas y facturas se filtren con base en:

cédula del usuario,
PDV actual,
sede actual,
empresa actual.

Con esto se evita que un usuario vea información de otra empresa, otra sede o otro PDV que no corresponda a su sesión activa.

5. Ajuste en Ventas Manuales para búsqueda y carga de productos
Se revisó la lógica de la venta manual para que la búsqueda de productos y el proceso de carga de pedidos queden alineados al entorno multiempresa, multisede y multipdv, usando el contexto del usuario autenticado en lugar de asumir la primera relación encontrada. La vista también quedó organizada para trabajar comercialmente con Valor Unidad y Valor Venta.

6. Ajuste en POS para respetar la sede y el PDV de la sesión
En el módulo POS se trabajó el servicio para que la pantalla, la búsqueda de productos y el proceso de facturación se basen en la sede activa y el PDV del usuario, tomando los datos del contexto actual en lugar de una asignación genérica. El POS también quedó usando ValorUnidad y ValorVenta como base comercial.

7. Resultado del ajuste
Con los cambios realizados hoy, quedó funcional la lógica para que Ventas Manuales y POS operen con la empresa actual del usuario, la sede actual y el PDV actual, evitando que el sistema tome la primera empresa disponible cuando el usuario tiene acceso a varias.

Conclusión
Se dejó implementada y validada la base de trabajo para que los módulos comerciales respeten correctamente el entorno multiempresa, multisede y multipdv, usando el contexto real de login del usuario como referencia principal.

1. Login multiempresa, multisede y multi-PDV

Se revisó el flujo actual de login, donde el usuario primero ingresa cédula y contraseña, luego selecciona un PDV y finalmente se crean los claims de sesión.

Se identificó que el sistema estaba tomando una sola sede con FirstOrDefault(), lo cual limitaba el acceso cuando un usuario tenía varias empresas o varias sedes asociadas.

Se propuso ajustar la lógica para que el usuario pueda entrar según este orden:

Usuario → Empresa → Sede → PDV

De esta forma, si un mismo usuario está asociado a varias compañías, puede escoger correctamente desde qué empresa, sede y punto de venta desea trabajar.

2. Selección de Empresa, Sede y PDV en el login

Se planteó cambiar la vista ValidarPDV para que ya no muestre directamente todos los PDV, sino que primero permita seleccionar la empresa, luego la sede correspondiente y finalmente el PDV disponible.

La información debe salir filtrada por el usuario conectado, no de forma general. Es decir, solo se deben mostrar empresas, sedes y PDV realmente configurados para esa cédula.

3. Ajuste de cargos por sede

También se identificó que el método ObtenerRolPermisos estaba tomando el primer cargo encontrado para la cédula, lo cual puede fallar si el usuario pertenece a varias empresas o sedes.

Se propuso crear una lógica como:

ObtenerRolPermisosPorSede(int cedula, int sedeId)

para que el cargo del usuario se obtenga según la sede seleccionada en el login.

4. Configuración de sede y cargo para empleados

Se revisó la vista SeleccionarSedeYCargo, usada cuando al buscar una cédula el sistema no encuentra configuración previa en sedeempleado.

Se propuso que esa vista ya no dependa de una sola empresa enviada por ViewBag.IdEmpresa, sino que consulte las empresas asociadas al empleado en EmpleadoEmpresa.

La nueva lógica quedó planteada así:

Cedula → EmpleadoEmpresa → Empresas asociadas → Sedes de esas empresas → Cargos de esa empresa

En el select de sedes se sugirió mostrar algo como:

Sede Principal - Empresa A
Sucursal Norte - Empresa B

Y al seleccionar una sede, el sistema debe filtrar los cargos que pertenecen a la empresa de esa sede.

5. Mejora del portal de inicio

Se habló de reorganizar la pantalla de inicio para que deje de ser una vista general y pase a ser un portal administrativo contextual.

La idea es que el inicio muestre información de la empresa, sede y PDV donde el usuario inició sesión, usando los claims creados durante el login:

EmpresaId
NombreEmpresa
SedeId
NombreSede
PdvId
NombrePdv
Cedula
Nombre
NombreRol

Así, si el usuario entra por una sede o empresa diferente, el portal cambia automáticamente según ese contexto.

6. Vista administrativa propuesta

Se propuso una vista tipo dashboard con:

Empresa conectada
Sede conectada
Punto de venta actual
Usuario
Cargo
Cantidad de sedes de la empresa
Cantidad de PDV de la sede
Cantidad de empleados en la sede
Accesos rápidos a módulos

Todo filtrado por el contexto del usuario, evitando consultar datos globales del sistema.

Resumen general

Lo logrado hasta ahora fue aterrizar la estructura para que el sistema funcione correctamente como:

Multiempresa + Multisede + MultiPDV

ya no solo visualmente, sino también desde la lógica de login, permisos, cargo, claims y portal inicial. Esto deja una base mucho más sólida para que cada usuario trabaje únicamente sobre la empresa, sede y punto de venta que seleccionó al ingresar.

+++ Login/Servicios/Implementacion/ReporteService.cs
@@ -498,7 +498,7 @@ namespace Plataforma.Servicios.Implementacion
                     FechaEmisionFactura = item.FechaEmisionFacturaReal ?? item.FechaEmisionFacturaVenta,
 
                     IdCliente = item.IdCliente,
-                    CedulaCliente = item.CedulaCliente,
+                    CedulaCliente = (int)item.CedulaCliente,
 
                     CedulaVendedor = item.Cedula,
